+++ src/Rule/Assertion/Declaration/ShouldBeFinal/ShouldBeFinal.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace PHPat\Rule\Assertion\Declaration\ShouldBeFinal;
 
 use PHPat\Configuration;
-use PHPat\Rule\Assertion\Declaration\Declaration;
 use PHPat\Rule\Assertion\Declaration\DeclarationAssertion;
 use PHPat\Rule\Assertion\Declaration\ValidationTrait;
 use PHPat\Statement\Builder\StatementBuilderFactory;
@@ -35,11 +34,6 @@ abstract class ShouldBeFinal extends DeclarationAssertion
     protected function applyValidation(ClassReflection $subject, bool $meetsDeclaration): array
     {
         return $this->applyShould($subject, $meetsDeclaration);
-    }
-
-    protected function getDeclarationType(): string
-    {
-        return Declaration::FINAL;
     }
 
     protected function getMessage(string $subject): string