+++ src/Rule/Assertion/Declaration/ShouldNotBeAbstract/ShouldNotBeAbstract.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PHPat\Rule\Assertion\Declaration\ShouldNotBeFinal;
+namespace PHPat\Rule\Assertion\Declaration\ShouldNotBeAbstract;
 
 use PHPat\Configuration;
 use PHPat\Rule\Assertion\Declaration\DeclarationAssertion;
@@ -12,7 +12,7 @@ use PHPStan\Reflection\ClassReflection;
 use PHPStan\Reflection\ReflectionProvider;
 use PHPStan\Type\FileTypeMapper;
 
-abstract class ShouldNotBeFinal extends DeclarationAssertion
+abstract class ShouldNotBeAbstract extends DeclarationAssertion
 {
     use ValidationTrait;
 
@@ -38,6 +38,6 @@ abstract class ShouldNotBeFinal extends DeclarationAssertion
 
     protected function getMessage(string $subject): string
     {
-        return sprintf('%s should not be final', $subject);
+        return sprintf('%s should not be abstract', $subject);
     }
 }