+++ src/Rule/Assertion/Relation/ShouldNotDepend/StaticMethodRule.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PHPat\Rule\Assertion\Relation\ShouldNotDepend;
+
+use PHPat\Rule\Extractor\Relation\StaticMethodCallExtractor;
+use PhpParser\Node;
+use PHPStan\Rules\Rule;
+
+/**
+ * @implements Rule<Node\Expr\StaticCall>
+ */
+class StaticMethodRule extends ShouldNotDepend implements Rule
+{
+    use StaticMethodCallExtractor;
+}