+++ src/Test/Builder/AssertionStep.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace PHPat\Test\Builder;
 
+use PHPat\Rule\Assertion\Declaration\ShouldBeAbstract\ShouldBeAbstract;
 use PHPat\Rule\Assertion\Declaration\ShouldBeFinal\ShouldBeFinal;
+use PHPat\Rule\Assertion\Declaration\ShouldNotBeAbstract\ShouldNotBeAbstract;
 use PHPat\Rule\Assertion\Declaration\ShouldNotBeFinal\ShouldNotBeFinal;
 use PHPat\Rule\Assertion\Relation\ShouldExtend\ShouldExtend;
 use PHPat\Rule\Assertion\Relation\ShouldImplement\ShouldImplement;
@@ -15,6 +17,20 @@ use PHPat\Rule\Assertion\Relation\ShouldNotImplement\ShouldNotImplement;
 
 class AssertionStep extends AbstractStep
 {
+    public function shouldBeAbstract(): Rule
+    {
+        $this->rule->assertion = ShouldBeAbstract::class;
+
+        return new BuildStep($this->rule);
+    }
+
+    public function shouldNotBeAbstract(): Rule
+    {
+        $this->rule->assertion = ShouldNotBeAbstract::class;
+
+        return new BuildStep($this->rule);
+    }
+
     public function shouldBeFinal(): Rule
     {
         $this->rule->assertion = ShouldBeFinal::class;