+++ tests/unit/rules/ShouldBeAbstract/AbstractClassTest.php
@@ -2,41 +2,40 @@
 
 declare(strict_types=1);
 
-namespace Tests\PHPat\unit\rules\ShouldNotDepend;
+namespace Tests\PHPat\unit\rules\ShouldBeAbstract;
 
 use PHPat\Configuration;
-use PHPat\Rule\Assertion\Relation\ShouldNotDepend\MethodParamRule;
-use PHPat\Rule\Assertion\Relation\ShouldNotDepend\ShouldNotDepend;
+use PHPat\Rule\Assertion\Declaration\ShouldBeAbstract\AbstractRule;
+use PHPat\Rule\Assertion\Declaration\ShouldBeAbstract\ShouldBeAbstract;
 use PHPat\Selector\Classname;
 use PHPat\Statement\Builder\StatementBuilderFactory;
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
 use PHPStan\Type\FileTypeMapper;
 use Tests\PHPat\fixtures\FixtureClass;
-use Tests\PHPat\fixtures\Simple\SimpleInterface;
 use Tests\PHPat\unit\FakeTestParser;
 
 /**
- * @extends RuleTestCase<MethodParamRule>
+ * @extends RuleTestCase<AbstractRule>
  */
-class MethodParamTest extends RuleTestCase
+class AbstractClassTest extends RuleTestCase
 {
     public function testRule(): void
     {
         $this->analyse(['tests/fixtures/FixtureClass.php'], [
-            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleInterface::class), 35],
+            [sprintf('%s should be abstract', FixtureClass::class), 29],
         ]);
     }
 
     protected function getRule(): Rule
     {
         $testParser = FakeTestParser::create(
-            ShouldNotDepend::class,
+            ShouldBeAbstract::class,
             [new Classname(FixtureClass::class, false)],
-            [new Classname(SimpleInterface::class, false)]
+            []
         );
 
-        return new MethodParamRule(
+        return new AbstractRule(
             new StatementBuilderFactory($testParser),
             new Configuration(false),
             $this->createReflectionProvider(),