+++ tests/unit/rules/ShouldNotBeAbstract/AbstractClassTest.php
@@ -2,41 +2,40 @@
 
 declare(strict_types=1);
 
-namespace Tests\PHPat\unit\rules\ShouldNotDepend;
+namespace Tests\PHPat\unit\rules\ShouldNotBeAbstract;
 
 use PHPat\Configuration;
-use PHPat\Rule\Assertion\Relation\ShouldNotDepend\NewRule;
-use PHPat\Rule\Assertion\Relation\ShouldNotDepend\ShouldNotDepend;
+use PHPat\Rule\Assertion\Declaration\ShouldNotBeAbstract\AbstractRule;
+use PHPat\Rule\Assertion\Declaration\ShouldNotBeAbstract\ShouldNotBeAbstract;
 use PHPat\Selector\Classname;
 use PHPat\Statement\Builder\StatementBuilderFactory;
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
 use PHPStan\Type\FileTypeMapper;
-use Tests\PHPat\fixtures\FixtureClass;
-use Tests\PHPat\fixtures\Simple\SimpleClass;
+use Tests\PHPat\fixtures\Simple\SimpleAbstractClass;
 use Tests\PHPat\unit\FakeTestParser;
 
 /**
- * @extends RuleTestCase<NewRule>
+ * @extends RuleTestCase<AbstractRule>
  */
-class NewTest extends RuleTestCase
+class AbstractClassTest extends RuleTestCase
 {
     public function testRule(): void
     {
-        $this->analyse(['tests/fixtures/FixtureClass.php'], [
-            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClass::class), 47],
+        $this->analyse(['tests/fixtures/Simple/SimpleAbstractClass.php'], [
+            [sprintf('%s should not be abstract', SimpleAbstractClass::class), 7],
         ]);
     }
 
     protected function getRule(): Rule
     {
         $testParser = FakeTestParser::create(
-            ShouldNotDepend::class,
-            [new Classname(FixtureClass::class, false)],
-            [new Classname(SimpleClass::class, false)]
+            ShouldNotBeAbstract::class,
+            [new Classname(SimpleAbstractClass::class, false)],
+            []
         );
 
-        return new NewRule(
+        return new AbstractRule(
             new StatementBuilderFactory($testParser),
             new Configuration(false),
             $this->createReflectionProvider(),