+++ tests/unit/rules/ShouldNotDepend/DocParamTagTest.php
@@ -33,14 +33,14 @@ class DocParamTagTest extends RuleTestCase
     public function testRule(): void
     {
         $this->analyse(['tests/fixtures/FixtureClass.php'], [
-            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClass::class), 64],
-            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClassTwo::class), 64],
-            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClassThree::class), 64],
-            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClassFour::class), 64],
-            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClassFive::class), 64],
-            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClassSix::class), 64],
-            [sprintf('%s should not depend on %s', FixtureClass::class, InterfaceWithTemplate::class), 64],
-            [sprintf('%s should not depend on %s', FixtureClass::class, ClassImplementing::class), 64],
+            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClass::class), 70],
+            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClassTwo::class), 70],
+            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClassThree::class), 70],
+            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClassFour::class), 70],
+            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClassFive::class), 70],
+            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleClassSix::class), 70],
+            [sprintf('%s should not depend on %s', FixtureClass::class, InterfaceWithTemplate::class), 70],
+            [sprintf('%s should not depend on %s', FixtureClass::class, ClassImplementing::class), 70],
         ]);
     }