+++ tests/unit/rules/ShouldNotDepend/StaticCallTest.php
@@ -5,26 +5,26 @@ declare(strict_types=1);
 namespace Tests\PHPat\unit\rules\ShouldNotDepend;
 
 use PHPat\Configuration;
-use PHPat\Rule\Assertion\Relation\ShouldNotDepend\MethodParamRule;
 use PHPat\Rule\Assertion\Relation\ShouldNotDepend\ShouldNotDepend;
+use PHPat\Rule\Assertion\Relation\ShouldNotDepend\StaticMethodRule;
 use PHPat\Selector\Classname;
 use PHPat\Statement\Builder\StatementBuilderFactory;
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
 use PHPStan\Type\FileTypeMapper;
 use Tests\PHPat\fixtures\FixtureClass;
-use Tests\PHPat\fixtures\Simple\SimpleInterface;
+use Tests\PHPat\fixtures\Special\ClassWithStaticMethod;
 use Tests\PHPat\unit\FakeTestParser;
 
 /**
- * @extends RuleTestCase<MethodParamRule>
+ * @extends RuleTestCase<StaticMethodRule>
  */
-class MethodParamTest extends RuleTestCase
+class StaticCallTest extends RuleTestCase
 {
     public function testRule(): void
     {
         $this->analyse(['tests/fixtures/FixtureClass.php'], [
-            [sprintf('%s should not depend on %s', FixtureClass::class, SimpleInterface::class), 35],
+            [sprintf('%s should not depend on %s', FixtureClass::class, ClassWithStaticMethod::class), 57],
         ]);
     }
 
@@ -33,10 +33,10 @@ class MethodParamTest extends RuleTestCase
         $testParser = FakeTestParser::create(
             ShouldNotDepend::class,
             [new Classname(FixtureClass::class, false)],
-            [new Classname(SimpleInterface::class, false)]
+            [new Classname(ClassWithStaticMethod::class, false)]
         );
 
-        return new MethodParamRule(
+        return new StaticMethodRule(
             new StatementBuilderFactory($testParser),
             new Configuration(false),
             $this->createReflectionProvider(),